refactor(data): style 분리
