+++ components/connect/Data/index.tsx
@@ -1,5 +1,4 @@
 import DropDown from 'components/common/DropDown';
-import styled from 'styled-components';
 import fileExist from 'public/assets/icons/connect/fileExists.svg';
 import supervisorCategoryOpen from 'public/assets/icons/connect/supervisorCategoryOpen.svg';
 import movePageL from 'public/assets/icons/connect/movePageL.svg';
@@ -7,177 +6,7 @@ import movePageR from 'public/assets/icons/connect/movePageR.svg';
 import Image from 'next/image';
 import DataDetail from './DataDetail';
 import { useState } from 'react';
-
-const DataBox = styled.div`
-  padding: 0 1.2rem 0.1rem;
-`;
-
-const SelectSupervisorBox = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 1.5rem;
-  text-align: center;
-  padding: 7.3rem 0 9.4rem;
-  cursor: pointer;
-
-  color: var(--color-black);
-  text-align: center;
-  font-size: 1.8rem;
-  font-weight: 600;
-  line-height: normal;
-`;
-
-const DataNavBox = styled.div`
-  display: flex;
-  align-items: center;
-  padding-bottom: 2rem;
-  border-bottom: 1px solid var(--color-gray);
-  position: relative;
-`;
-
-const DataNav = styled.nav`
-  color: #979797;
-  text-align: center;
-  font-size: 1.8rem;
-  font-weight: 600;
-  line-height: normal;
-
-  ul {
-    display: flex;
-    align-items: center;
-  }
-
-  li {
-    width: 9.2rem;
-
-    &:not(:last-child) {
-      margin-right: 6rem;
-    }
-  }
-`;
-
-const DataNavInputBox = styled.div`
-  margin-left: auto;
-  display: flex;
-
-  input {
-    width: 43.6rem;
-    height: 3.5rem;
-    background-color: var(--color-white);
-    margin-left: -16.3rem;
-    padding-left: 18.3rem;
-    border-radius: 10rem;
-  }
-`;
-
-const DataDropDownBox = styled.div`
-  position: absolute;
-  top: 0;
-  right: 27.3rem;
-`;
-
-const Label = styled.div`
-  padding: 1rem 0;
-  display: flex;
-  align-items: center;
-  text-align: center;
-
-  .label__1 {
-    width: 9.2rem;
-  }
-
-  .label__2 {
-    padding-left: 1.7rem;
-  }
-
-  .label__3 {
-    margin-left: auto;
-    width: 12.7rem;
-  }
-
-  .label__4 {
-    padding: 0 5.5rem 0 3.6rem;
-  }
-
-  color: var(--color-black);
-  font-size: 1.5rem;
-  font-weight: 500;
-  line-height: normal;
-`;
-
-const LabelBox = styled(Label)`
-  border-bottom: 1px solid var(--color-gray);
-  margin-bottom: 0.1rem;
-`;
-
-const DataListBox = styled(Label)`
-  padding: 3.3rem 0;
-  cursor: pointer;
-
-  &:hover {
-    background-color: #f7f7f9;
-  }
-
-  div {
-    display: flex;
-    align-items: center;
-    gap: 2rem;
-  }
-
-  .label__1 {
-    font-weight: 600;
-  }
-
-  .label__2 {
-  }
-
-  .label__3 {
-    color: #979797;
-  }
-`;
-
-const PaginationBox = styled.div`
-  display: flex;
-  align-items: center;
-  gap: 1.5rem;
-  margin: 13rem auto;
-  width: 42.1rem;
-
-  color: var(--color-black);
-  text-align: center;
-  font-size: 2.2rem;
-  font-weight: 500;
-  line-height: normal;
-
-  p,
-  img {
-    cursor: pointer;
-  }
-
-  p {
-    width: 5rem;
-    height: 3.2rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-
-  img:first-child {
-    margin-right: auto;
-  }
-
-  img:last-child {
-    margin-left: auto;
-  }
-
-  .pagination__active {
-    color: var(--color-white);
-    font-size: 2.2rem;
-    background-color: var(--color-black);
-    border-radius: 10rem;
-  }
-`;
+import * as S from './style';
 
 /** connect - 자료 페이지 */
 const Data = () => {
@@ -186,8 +15,8 @@ const Data = () => {
   return (
     <>
       {temp && (
-        <DataBox>
-          <SelectSupervisorBox>
+        <S.DataBox>
+          <S.SelectSupervisorBox>
             <p>이혜리 슈퍼바이저</p>
             <Image
               src={supervisorCategoryOpen}
@@ -195,18 +24,18 @@ const Data = () => {
               width={8}
               height={7}
             />
-          </SelectSupervisorBox>
-          <DataNavBox>
-            <DataNav>
+          </S.SelectSupervisorBox>
+          <S.DataNavBox>
+            <S.DataNav>
               <ul>
                 <li>공지사항</li>
                 <li>레시피</li>
                 <li>행사내용</li>
                 <li>기타</li>
               </ul>
-            </DataNav>
-            <DataNavInputBox>
-              <DataDropDownBox>
+            </S.DataNav>
+            <S.DataNavInputBox>
+              <S.DataDropDownBox>
                 <DropDown
                   width={16.3}
                   height={3.5}
@@ -215,18 +44,18 @@ const Data = () => {
                   toggleTopSize={50}
                   list={['글 제목', '글 내용']}
                 />
-              </DataDropDownBox>
+              </S.DataDropDownBox>
               <input type="text" />
-            </DataNavInputBox>
-          </DataNavBox>
-          <LabelBox>
+            </S.DataNavInputBox>
+          </S.DataNavBox>
+          <S.LabelBox>
             <p className="label__1">번호</p>
             <p className="label__2">제목</p>
             <p className="label__3">작성일자</p>
             <p className="label__4">이해리 슈퍼바이저</p>
-          </LabelBox>
+          </S.LabelBox>
           <div>
-            <DataListBox>
+            <S.DataListBox>
               <p className="label__1">06</p>
               <div className="label__2">
                 <p>알바생 교육관련 6월 공지사항</p>
@@ -234,8 +63,8 @@ const Data = () => {
               </div>
               <p className="label__3">2023.06.13</p>
               <p className="label__4">이혜리 슈퍼바이저</p>
-            </DataListBox>
-            <DataListBox
+            </S.DataListBox>
+            <S.DataListBox
               onClick={() => {
                 setTemp(false);
               }}
@@ -247,8 +76,8 @@ const Data = () => {
               </div>
               <p className="label__3">2023.05.13</p>
               <p className="label__4">이혜리 슈퍼바이저</p>
-            </DataListBox>
-            <DataListBox>
+            </S.DataListBox>
+            <S.DataListBox>
               <p className="label__1">04</p>
               <div className="label__2">
                 <p>알바생 교육관련 4월 공지사항</p>
@@ -256,8 +85,8 @@ const Data = () => {
               </div>
               <p className="label__3">2023.04.13</p>
               <p className="label__4">이혜리 슈퍼바이저</p>
-            </DataListBox>
-            <DataListBox>
+            </S.DataListBox>
+            <S.DataListBox>
               <p className="label__1">03</p>
               <div className="label__2">
                 <p>알바생 교육관련 3월 공지사항</p>
@@ -265,8 +94,8 @@ const Data = () => {
               </div>
               <p className="label__3">2023.03.13</p>
               <p className="label__4">이혜리 슈퍼바이저</p>
-            </DataListBox>
-            <DataListBox>
+            </S.DataListBox>
+            <S.DataListBox>
               <p className="label__1">02</p>
               <div className="label__2">
                 <p>알바생 교육관련 2월 공지사항</p>
@@ -274,8 +103,8 @@ const Data = () => {
               </div>
               <p className="label__3">2023.02.13</p>
               <p className="label__4">이혜리 슈퍼바이저</p>
-            </DataListBox>
-            <DataListBox>
+            </S.DataListBox>
+            <S.DataListBox>
               <p className="label__1">01</p>
               <div className="label__2">
                 <p>알바생 교육관련 1월 공지사항</p>
@@ -283,9 +112,9 @@ const Data = () => {
               </div>
               <p className="label__3">2023.01.13</p>
               <p className="label__4">이혜리 슈퍼바이저</p>
-            </DataListBox>
+            </S.DataListBox>
           </div>
-          <PaginationBox>
+          <S.PaginationBox>
             <Image
               src={movePageL}
               alt="pagination-move-left"
@@ -303,8 +132,8 @@ const Data = () => {
               width={16}
               height={9}
             />
-          </PaginationBox>
-        </DataBox>
+          </S.PaginationBox>
+        </S.DataBox>
       )}
       {!temp && <DataDetail />}
     </>
